requirements.txt
thordata-sdk>=1.8.4
langchain>=0.3.0
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
chromadb>=0.4.0
markdownify>=0.11.6
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
html2text>=2024.2.26
beautifulsoup4>=4.12.0
lxml>=5.0.0
tiktoken>=0.5.0
openai>=1.0.0